+++ src/app/_lib/types/songs-interface.ts
@@ -1,0 +1,11 @@
+export interface Songs {
+    id: number;
+    artist: string;
+    track: string;
+    image: string;
+    video: string;
+    streams: {
+      monthly: number;
+      total: number;
+    };
+  }
